add seek and play time

+++ ios/AdioManager+URLSessionDelegates.swift
@@ -9,18 +9,25 @@
 import Foundation
 
 extension AdioManager: URLSessionDownloadDelegate {
+  // If the downloadSession hasn't been created yet we need to create it
+  func setDownloadSession() {
+    if downloadManager.downloadSession == nil {
+      downloadManager.downloadSession = downloadSession
+    }
+  }
+  
+  func downloadClip(_ clip: Clip) {
+    setDownloadSession()
+    downloadManager.startDownload(clip)
+  }
+  
   func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didFinishDownloadingTo location: URL) {
-    
-    
     guard let src = downloadTask.originalRequest?.url else { return }
     
-//    let download = downloadManager.activeDownloads[src]
-//    downloadManager.activeDownloads[src] = nil
+    guard let download = downloadManager.activeDownloads.removeValue(forKey: src) else { return }
     
-    let download = downloadManager.activeDownloads.removeValue(forKey: src)
-    
-    let destinationURL = download?.clip.localSrc
-    
+    let destinationURL = download.clips[0].localSrc
+
     let fileManager = FileManager.default
     
     try? fileManager.removeItem(at: destinationURL!)
@@ -28,10 +35,14 @@ extension AdioManager: URLSessionDownloadDelegate {
     do {
       try fileManager.moveItem(at: location, to: destinationURL!)
       
-//      download?.clip.isDownloaded = true
-      download?.clip.setIsDownloaded()
+      for clip in download.clips {
+        clip.setIsDownloaded()
+        
+        audioSession?.addClip(clip)
+        
+        self.sendEvent(withName: AdioManagerConstants.Events.ClipDownloaded, body: clip.data())
+      }
       
-      self.sendEvent(withName: AdioManagerConstants.Events.ClipDownloaded, body: download?.clip.data())
     } catch let error {
       print("Could not copy file to disk: \(error.localizedDescription)")
     }
@@ -46,10 +57,12 @@ extension AdioManager: URLSessionDownloadDelegate {
     
     download?.progress = progress
     
-    let clipId = download?.clip.id
+    let clips = download?.clips
     
-    self.sendEvent(withName: AdioManagerConstants.Events.ClipDownloadProgress, body: [
-      "id": clipId!,
-      "progress": progress])
+    for clip in clips! {
+      self.sendEvent(withName: AdioManagerConstants.Events.ClipDownloadProgress, body: [
+        "id": clip.id,
+        "progress": progress])
+    }
   }
 }
